Updating security config

+++ src/main/java/com/example/demo/security/JwtTokenHelper.java
@@ -8,7 +8,6 @@ import java.util.Map;
 import javax.crypto.SecretKey;
 
 import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
@@ -19,7 +18,7 @@ import io.jsonwebtoken.security.Keys;
 public class JwtTokenHelper implements Serializable {
 
 	private static final long serialVersionUID = 7008375124389347049L;
-	public static final long TOKEN_VALIDITY = 1 * 60 * 60;
+	public static final long TOKEN_VALIDITY = (long)1 * 60 * 60;
 	private static final SecretKey KEY = Keys.secretKeyFor(SignatureAlgorithm.HS256); //or HS384 or HS512;
 
 	public String generateJwtToken(UserDetails userDetails) {
